Fit a Simple Numerosity Model

This model assumes the distribution of mental representation for a given number/count k is N(k, (w * k) ^ 2).

Usage

fit_numerosity(data, name_bigset, name_smallset, name_acc, n_fit = 5, seed = 1)

Arguments

data

Raw data of class data.frame.

name_bigset, name_smallset

Variable name in data indicates bigger and smaller set.

name_acc

Variable name in data indicates user's response is correct or not.

n_fit

Number of fits to try to find the best estimate.

seed

Random seed. Default is 1 so that results can be reproduced.

Value

A list() with structure the same as optim().