Multiple Sensory Integration

There will typically be some speed advantage if there are more than one sensory inputs to be employed. This function calculates this advantage.

Usage

multisense(data, .by = NULL, .input = NULL, .extra = NULL)

Arguments

data

Raw data of class data.frame.

.by

The column name(s) in data used to be grouped by. If set to NULL (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.

.input, .extra

Each is a list() containing all the input variable names and special values for certain variables. See more in the details section.

Value

An object with the same class as data contains following values:

mrt_image

Mean reaction time of Image stimuli.

mrt_sound

Mean reaction time of Sound stimuli.

mrt_mixed

Mean reaction time of Mixed stimuli.

mrt_mixadv

Mean reaction decrease of Mixed stimuli compared to other two types of stimuli.