N Back Paradigm

A classical working memory test.

Usage

nback(data, .by = NULL, .input = NULL, .extra = NULL)

dualnback(data, .by = NULL, .input = NULL, .extra = NULL)

Arguments

data

Raw data of class data.frame.

.by

The column name(s) in data used to be grouped by. If set to NULL (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.

.input, .extra

Each is a list() containing all the input variable names and special values for certain variables. See more in the details section.

Value

An object with the same class as data contains following values (tripled for dual n-back):

pc

Percentage of correct responses.

mrt

Mean reaction time.

ies

Inverse efficiency score.

rcs

Rate correct score.

lisas

Linear integrated speed-accuracy score.

dprime

Sensitivity index.

Details

Note for dual n-back, the returned indices include those for each modal and both modals so they are tripled with additional suffix after each index name.