Task Switching and Stroop-like paradigm

In task switching paradigms, two types of tasks switch between each other, so the "switch cost" can be calculated (using switchcost()). Similarly, in Stroop-like tasks, stimuli are classified into two conditions (i.e., "congruent" and "incongruent"), so the "congruence effect" can be calculated (using congeff()). There are also special types of tests where congruence effect and switch cost both exist, from which complexswitch() calculates both.

Usage

complexswitch(data, .by = NULL, .input = NULL, .extra = NULL)

congeff(data, .by = NULL, .input = NULL, .extra = NULL)

switchcost(data, .by = NULL, .input = NULL, .extra = NULL)

Arguments

data

Raw data of class data.frame.

.by

The column name(s) in data used to be grouped by. If set to NULL (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.

.input, .extra

Each is a list() containing all the input variable names and special values for certain variables. See more in the details section.

Value

A tibble with the following variables:

For the total task:

pc

Percent of correct.

mrt

Mean reaction time.

For congruence effect and switch cost, the following indices will be included (including diffs and value for each condition):

pc

Percent of correct.

mrt

Mean reaction time.

ies

Inverse efficiency score.

rcs

Rate correct score.

lisas

Linear integrated speed-accuracy score.