Calculate basic speed and accuracy scores

This function calculates reaction time and accuracy scores, which are very basic to most tests.

calc_spd_acc(
  data,
  .by,
  name_acc,
  name_rt,
  rm.out = TRUE,
  rt_rtn = c("both", "mean", "sd", "none"),
  acc_rtn = c("both", "count", "percent", "none")
)

Arguments

data

Raw data of class data.frame.

.by

The column name(s) in data used to be grouped by. If set to NULL (default), all data will be treated as from one subject and there will be no grouping columns in the value returned.

name_acc

The column name of the data input whose values are user's correctness, in which is a numeric vector so coded that 1 means scoring correct, 0 means scoring incorrect, and that -1 means no response is made.

name_rt

The column name of the data input whose values are user's reaction time, in which is a numeric vector with NA values representing abnormal reaction times.

rm.out

A logical value. Whether remove outlier or not. Default to TRUE.

rt_rtn

If mean or standard deviation of response times should be returned.

acc_rtn

If count or percent of correct responses should be returned.

Value

A tibble contains the required scores.

Details

A major part of behavior tests is the speed and accuracy metrics, which are actually the only ones that any test collects. Based on these two metrics, two basic groups of scores can be obtained. The first group contains the mean and standard deviations of response times the number, and the second contains the number and percent of correct responses.

Note

Outlier removal is recommended and it is set as default in here. The outlier detection method used here is the "interquantile range" rule suggested by Tukey (1977).