Outliers Detection for response time data

Outliers Detection for response time data

Usage

check_outliers_rt(
  x,
  method = c("transform", "z_score", "cutoff"),
  threshold = NULL
)

Arguments

x

A vector of input reaction time data.

method

The method used to detect outliers. If set to "transform", a square root transformation is applied to the data before applying "z_score" method outlier detection, see Cousineau & Chartier (2010). If set to "z_score", any value with absolute z-score larger than threshold is considered as outlier. If set to "cutoff", the any value out of threshold range is considered as outlier.

threshold

The threshold for determining whether a value is outlier or not. For "transform" and "z_score" method, the default is 2.5. For "cutoff" method, the default is c(0.2, Inf).

Value

A logical vector of the detected outliers.